+++ src/Contracts/SettingInterface.php
@@ -9,4 +9,13 @@ interface SettingInterface
     public function getKey(): string;
 
     public function getDefaultValue(): mixed;
+
+    public function getLabel(): string;
+
+    public function getFieldType(): string;
+
+    /**
+     * @return string[]
+     */
+    public function getOptions(): array;
 }